Bug 547195 - [Cocoa] Clean up .bridgesupport.extras
* Remove all uses of swt_alloc, swt_param_name, swt_param_cast,
  swt_java_type64.
* Remove redundant uses of swt_java_type.
* Fix incorrect swt_java_type in CGContextSetLineDash.
* Migrate from obsolete protocol names to modern equivalents.

+++ bundles/org.eclipse.swt/Eclipse SWT/cocoa/org/eclipse/swt/graphics/GC.java
@@ -3075,7 +3075,7 @@ void initCGContext(long cgContext) {
 			case SWT.LINE_CUSTOM: dashes = data.lineDashes; break;
 		}
 		if (dashes != null) {
-			float[] lengths = new float[dashes.length];
+			double[] lengths = new double[dashes.length];
 			for (int i = 0; i < lengths.length; i++) {
 				lengths[i] = width == 0 || data.lineStyle == SWT.LINE_CUSTOM ? dashes[i] : dashes[i] * width;
 			}
